# backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Syspro APS Scheduler API
  version: "1.0.0"
  description: |
    Finite-capacity production scheduler that reads master data from SYSPRO
    and writes approved schedules back via the APS compatibility layer.
  contact:
    name: APS Development Team
servers:
  - url: http://localhost:3000/api
    description: Local development / production (same port)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        jobId:       { type: string }
        itemCode:    { type: string }
        description: { type: string }
        quantity:    { type: number }
        priority:    { type: integer }
        status:      { type: string }
        dueDate:     { type: string, format: date-time }
        releaseDate: { type: string, format: date-time }
        operations:  { type: array, items: { $ref: '#/components/schemas/Operation' } }
    Operation:
      type: object
      properties:
        opId:          { type: string }
        sequence:      { type: integer }
        description:   { type: string }
        workcentreId:  { type: string }
        setupTime:     { type: number, description: Minutes }
        runTime:       { type: number, description: Minutes per unit }
        quantity:      { type: number }
        status:        { type: string }

    # ── Resources ─────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        error:   { type: string }
        code:    { type: string }
        details: { type: object }
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServiceUnavailable:
      description: Database not connected
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:
...
  /jobs:
    get:
      summary: List open production jobs from SYSPRO
      tags: [Jobs]
      responses:
        "200":
          description: Array of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs: { type: array, items: { $ref: '#/components/schemas/Job' } }
        "503":
          $ref: '#/components/responses/ServiceUnavailable'
  /jobs/{jobId}:
    get:
      summary: Get a single job with its operations
      tags: [Jobs]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Job with operations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        "404":
          $ref: '#/components/responses/NotFound'
  /jobs/material-plan:
    post:
      summary: Compute material availability for a list of jobs
      tags: [Jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobIds: { type: array, items: { type: string } }
      responses:
        "200":
          description: Material plan per job
  /jobs/{jobId}/bom-detail:
    get:
      summary: Full BOM + shortage detail for a single job
      tags: [Jobs]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: BOM lines with availability

  # Resources
...